Fix checkstyle warnings
