+++ test/src/main/java/se/fortnox/reactivewizard/test/AbstractPojo.java
@@ -1,0 +1,4 @@
+package se.fortnox.reactivewizard.test;
+
+abstract class AbstractPojo {
+}
