+++ test/src/main/java/se/fortnox/reactivewizard/test/PrimitivePojo.java
@@ -1,0 +1,29 @@
+package se.fortnox.reactivewizard.test;
+
+class PrimitivePojo {
+    private float floatPrimitive;
+    private int intPrimitive;
+    private long longPrimitive;
+    private double doublePrimitive;
+    private boolean booleanPrimitive;
+
+    public float getFloatPrimitive() {
+        return floatPrimitive;
+    }
+
+    public int getIntPrimitive() {
+        return intPrimitive;
+    }
+
+    public long getLongPrimitive() {
+        return longPrimitive;
+    }
+
+    public double getDoublePrimitive() {
+        return doublePrimitive;
+    }
+
+    public boolean isBooleanPrimitive() {
+        return booleanPrimitive;
+    }
+}
